requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-active-link==0.1.8
django-auto-logout==0.5.0
django-extensions==3.2.0
django-filter==21.1
django-formtools==2.4
django-htmx==1.14.0
django-htmx-autocomplete==0.6.2
django-import-export==2.7.1
django-jalali-date==1.0.1
django-multiselectfield==0.1.12
djangorestframework==3.13.1
fastcore==1.5.28
hyperlink==21.0.0
importlib-metadata==4.11.3
incremental==22.10.0
jdatetime==4.1.0
Markdown==3.3.6
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==23.0
Pillow==9.4.0
service-identity==21.1.0
sweetify==2.3.1
Werkzeug==2.2.1
wfastcgi==3.0.0
zipp==3.8.0
